"udps" and "pudp" for datastreams; always ACK
It made sense to enable the data-stream processing (separating data streams
by logical properties) when reading VDIF frames from any UDP mechanism.

The "udps" protocol, combined with defined data streams now behaves as
"udpsnor" with data streams; the sequence numbers get stripped and counted
and statistics will be kept per sender but no effort will be made to
re-order the packets in each data stream (how could it?)

Also found that the back-traffic sending was disabled if code used the
per-sender::handle_seqnr(...). This back-traffic is very important for any
non-directly connected transfers so it was re-enabled.

There is now some logic in the per-sender::handle_seqnr(...) to prevent
doing statistics on something that doesn't look like a running sequence of
packet sequence numbers (e.g. the "plain UDP" data stream reader calls that
one with the same sequence number each time for there isn't any, but it
reuses the code to make the back-traffic flow)